[kvx] FPU vector builtins support improvement
Summary:
The following builtins have been reworked/added to sync with GCC support:

__builtin_kvx_cmovehq, __builtin_kvx_cmoveho, __builtin_kvx_cmovehx,
__builtin_kvx_cmovewp, __builtin_kvx_cmovewq, __builtin_kvx_cmovewo,
__builtin_kvx_cmovedp, __builtin_kvx_cmovedq, __builtin_kvx_cmovefwp,
__builtin_kvx_cmovefwq, __builtin_kvx_cmovefwo,
__builtin_kvx_cmovefdp, __builtin_kvx_cmovefdq, __builtin_kvx_fabswq,
__builtin_kvx_fabswo, __builtin_kvx_fabsdp, __buildin_kvx_fabsdq,
__builtin_kvx_fnegwq, __builtin_kvx_fnegwo, __builtin_kvx_fnegdp,
__builtin_kvx_fnegdq, __builtin_kvx_fmaxwq, __builtin_kvx_fmaxwo,
__builtin_kvx_fmaxdp, __builtin_kvx_fmaxdq, __builtin_kvx_fmindp,
__builtin_kvx_fmindq, __builtin_kvx_fminwp, __builtin_kvx_fminwo,
__builtin_kvx_fmindp, __builtin_kvx_fmindq, __builtin_kvx_fmm212w,
__builtin_kvx_fmma212w, __builtin_kvx_fmms212w, __builtin_kvx_faddwo,
__builtin_kvx_fadddq, __builtin_kvx_faddcwcq, __builtin_kvx_faddcdcp,
__builtin_kvx_fsbfwo, __builtin_kvx_fsbfdq, __builtin_kvx_fsbfcwcq,
__builtin_kvx_fsbfcdcp, __builtin_kvx_fmulwo, __builtin_kvx_fmuldq,
__builtin_kvx_fmulwcq, __builtin_kvx_fmulcwcq, __builtin_kvx_fmuldcp,
__builtin_kvx_fmulcdcp, __builtin_kvx_ffmawo, __builtin_kvx_ffmadq,
__builtin_kvx_ffmswo, __builtin_kvx_ffmsdq, __builtin_kvx_floatwp,
__builtin_kvx_floatwq, __builtin_kvx_floatwo, __builtin_kvx_floatdp,
__builtin_kvx_floatdq, __builtin_kvx_floatuwp, __builtin_kvx_floatuwq,
__builtin_kvx_floatuwo, __builtin_kvx_floatudp,
__builtin_kvx_floatudq, __builtin_kvx_fixedwp, __builtin_kvx_fixedwq,
__builtin_kvx_fixedwo, __builtin_kvx_fixeddp, __builtin_kvx_fixeddq,
__builtin_kvx_fixeduwp, __builtin_kvx_fixeduwq,
__builtin_kvx_fixeduwo, __builtin_kvx_fixedudp,
__builtin_kvx_fixedudq.

ref T12105

Test Plan: New codegen tests.

Reviewers: O81 llvm-coolidge!

Maniphest Tasks: T12105

Differential Revision: https://phab.kalray.eu/D8576

+++ llvm/test/CodeGen/KVX/builtins-scalar.c
@@ -360,63 +360,63 @@ unsigned long fixedud(double x) {
 float fcdivw(float v1, float v2) {
   return __builtin_kvx_fcdivw(v1, v2);
   // CHECK-LABEL: fcdivw:
-  // CHECK: fcdivw
+  // CHECK: fcdivw $r0 = $r0r1
   // CHECK: ret
 }
 
 double fcdivd(double v1, double v2) {
   return __builtin_kvx_fcdivd(v1, v2);
   // CHECK-LABEL: fcdivd:
-  // CHECK: fcdivd
+  // CHECK: fcdivd $r0 = $r0r1
   // CHECK: ret
 }
 
 float fsdivw(float v1, float v2) {
   return __builtin_kvx_fsdivw(v1, v2);
   // CHECK-LABEL: fsdivw:
-  // CHECK: fsdivw
+  // CHECK: fsdivw $r0 = $r0r1
   // CHECK: ret
 }
 
 double fsdivd(double v1, double v2) {
   return __builtin_kvx_fsdivd(v1, v2);
   // CHECK-LABEL: fsdivd:
-  // CHECK: fsdivd
+  // CHECK: fsdivd $r0 = $r0r1
   // CHECK: ret
 }
 
-float fsrecw(double v) {
+float fsrecw(float v) {
   return __builtin_kvx_fsrecw(v);
   // CHECK-LABEL: fsrecw:
-  // CHECK: fsrecw
+  // CHECK: fsrecw $r0 = $r0
   // CHECK: ret
 }
 
 double fsrecd(double v) {
   return __builtin_kvx_fsrecd(v);
   // CHECK-LABEL: fsrecd:
-  // CHECK: fsrecd
+  // CHECK: fsrecd $r0 = $r0
   // CHECK: ret
 }
 
 long sbmm8(long a, long b) {
   return __builtin_kvx_sbmm8(a, b);
   // CHECK-LABEL: sbmm8:
-  // CHECK: sbmm8
+  // CHECK: sbmm8 $r0 = $r0, $r1
   // CHECK: ret
 }
 
 long sbmmt8(long a, long b) {
   return __builtin_kvx_sbmmt8(a, b);
   // CHECK-LABEL: sbmmt8:
-  // CHECK: sbmmt8
+  // CHECK: sbmmt8 $r0 = $r0, $r1
   // CHECK: ret
 }
 
 long satd(long v, unsigned char b) {
   return __builtin_kvx_satd(v, b);
   // CHECK-LABEL: satd:
-  // CHECK: satd
+  // CHECK: satd $r0 = $r0, $r1
   // CHECK: ret
 }
